refactor(memory): replace the runtime pool with static storage
- Every consumer allocated its buffer once at startup and never freed
  it: 7 DevPool allocations, 0 frees, and destroyTask() had no callers,
  so the free path was unreachable at runtime
- The pool therefore carried ~600 lines of runtime bookkeeping over what
  is a fixed partition; the buffers become plain statics and the pool,
  its allocator and the devmem layer are removed
- Finish the reference cleanup the pool teardown implies: bus.h,
  testcmds, testsys and protocol.toml no longer name the pool
- Widen the DMA and impl hook lengths to 32 bits, which the UART
  override already assumed

Model: deepseek-flash DeepSeek
Agent: 白灵翎 (Hermes Agent)

+++ platform/STM32/peripherals/bus_spi.cpp
@@ -493,7 +493,7 @@ Result SpiBus::spiInternalReadWriteBufPolled(const uint8_t *txData,
 
 Result SpiBus::transmitReceiveImpl(TransferCallback cb, void *ctx,
                              const uint8_t *txData, uint8_t *rxData,
-                             uint16_t len) {
+                             uint32_t len) {
   if (len == 0) return Result::InvalidParam;
   if (!_initialized) return Result::NotReady;
