+++ src/drivers/peripherals/bus/bus_uart.h
@@ -82,7 +82,7 @@ private:
 
   Result transmitReceiveImpl(TransferCallback cb, void *ctx,
                         const uint8_t *txData, uint8_t *rxData,
-                        uint16_t len) override;
+                        uint32_t len) override;
 
 public:
   static constexpr uint32_t MAX_BUF_SIZE = _bufSize;
